refactor(api): move attachments to modules/ as controller, model, service
