+++ apps/api/src/modules/attachments/__tests__/integration/attachments.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from 'bun:test';
-import { api, authedApi } from '../../../__tests__/helpers/app';
-import { signUpTestUser } from '../../../__tests__/helpers/auth';
-import { resetDb } from '../../../__tests__/helpers/db';
+import { api, authedApi } from '#tests/helpers/app';
+import { signUpTestUser } from '#tests/helpers/auth';
+import { resetDb } from '#tests/helpers/db';
 
 // Attachments feature: metadata in Postgres, bytes in the object store (shared/
 // s3.ts against a real MinIO — see the Tests setup for S3_* env). This is the
