+++ apps/api/src/modules/attachments/service.ts
@@ -1,9 +1,9 @@
 import { db, issue, issueAttachment, issueFieldValue } from '@repo/db';
 import { eq, sql } from 'drizzle-orm';
-import { iso, num } from '../shared/lib';
+import { iso, num } from '#shared/lib';
 
 // Data access for issue attachments. File bytes live in the S3-compatible object
-// store (../shared/s3.ts); these rows hold the metadata and the object key. publicId is
+// store (#shared/s3); these rows hold the metadata and the object key. publicId is
 // the unguessable id used in the public download URL.
 
 export interface AttachmentRow {
